fix non-monotonically increasing frame_numbers
if a store was written with non-monotically increasing
framenumbers, such as when synchronized while recording,
the frame-numbers would wrap around zero; e.g. 6055, 6056, 0, 1, 2
and we would fail in get_next_image when wrapping.

the internal _get_next_framenumber_and_chunk_frame_idx was
only used in the 'read_next()' codepath, and so we could
implement a more direct check which actually checks if the
next chunk exists based on index